Fix proving key serialization (#212)
Recall that an rank 1 constraint has the form A * B = C.

Proving key serialization was incorrect because the "generate keypair" function in libsnark swaps all A and B terms in a constraint system if a certain condition is met. We were not doing this when recreating the constraint system ourselves, which led to creating bad proofs. I have corrected this issue and added a test which failed previously and now passes. Thanks to @deepthiskumar for help debugging.