Builds are portable (#252)
Tested on macOS, we may need to make more changes to the linux build
later. The biggest change here is explicitly setting `OPT_FLAGS` to not
include `-march=native -mtune=native`, and explicitly turning on
USE_ASM.

I ran transaction-snark-profiler over in Coda both with and without the
OPT_FLAGS override and there is negligible changes in performance
numbers. USE_ASM is what really gets this thing flying.

PERFORMANCE is disable always on macOS because there are errors in the
rest of the build that were always present when it was on.