Add Typ.t's for non-snarkable values
This will let us avoid the hacks that we used in Meja during the
conference to handle string passing etc. between the worlds.

This slightly overloads the behaviour of As_prover.Ref.t, but for Meja's
purposes this will be invisible to the user anyway. Perhaps more
importantly, this type can be special-cased to give good error messages
when the user tries to use prover-generated data outside of a prover
block.

Generating code that uses `snarkable` will be slightly more awkward
since we need to track when the same variable is written and then
stored, but it should be possible to make it work.  The important case
that this supports is code like:

```reason
...
let x = {data: 1000f, aux_data: 50};
let x = Prover { {x.. data= x.data * 10f} };
...
```

where the int [aux_data] might be used later in checked code for the
length of a loop etc.