Add Typ.Internal.fn for converting functions in the imperative API
This `Typ.t` describes how to convert between checked functions and
prover functions, using a `Typ.t` for each of the argument and the
return type.

The implementation is 'unsafe', but when used with the API in Snark_intf
it should function correctly. There is an additional check that we are
in prover mode, so that explicitly calling [Typ_monads.Read.read] fails
instead of generating a witness for the incorrect R1CS.