Use a seperate counter for Ident.fresh name suffixes
This ensures that the contents of the local dune `_build` directory, the
snarky opam directory, the `meji` directory, and the OCaml stdlib
directoy do not affect the names of fresh variables. Each .mli or .meji
file currently generates a new `Ident.t`, which changes the main Ident.t
counter; using this distinct counter isolates the names from these
effects.