[sponge] parameterize 'initial_ark' in poseidon permutation
Our implementation of poseidon has an optional round constant in the very beginning.
We use it thorough the OCaml codebase, but our new params do not use it.
See https://github.com/o1-labs/marlin/blob/master/oracle/src/poseidon.rs#L82

For this reason, we need to be able to disable when needed.